+++ opal/caregivers/factories.py
@@ -1,7 +1,9 @@
 """Module providing model factories for caregiver app models."""
 
-from factory import Sequence, SubFactory
+from factory import Faker, Sequence, SubFactory
 from factory.django import DjangoModelFactory
+from faker.generator import random
+from faker.providers import BaseProvider
 
 from opal.users.factories import Caregiver
 
@@ -37,6 +39,36 @@ class SecurityAnswer(DjangoModelFactory):
     question = 'Apple'
     user = SubFactory(CaregiverProfile)
     answer = 'answer'
+
+
+class DeviceProvider(BaseProvider):
+    """Faker Provider class that generates random values for the Device factory."""
+
+    def device_id(self) -> str:
+        """
+        Generate a random device_id.
+
+        Returns:
+            A random device_id value between 16 and 100 characters in length.
+        """
+        length = random.randint(16, 100)
+        char_choices = [str(digit) for digit in range(9)] + ['a', 'b', 'c', 'd', 'e', 'f']  # 0-9 digits and letters up to 'f' # noqa: WPS221, E501
+        chars = [random.choice(char_choices) for _ in range(length)]
+        return ''.join(chars)
+
+
+Faker.add_provider(DeviceProvider)
+
+
+class Device(DjangoModelFactory):
+    """Model factory to create [opal.caregivers.models.Device][] models."""
+
+    class Meta:
+        model = models.Device
+
+    caregiver = SubFactory(CaregiverProfile)
+    type = models.DeviceType.IOS  # noqa: A003
+    device_id = Faker('device_id')
 
 
 class RegistrationCode(DjangoModelFactory):