+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Add logo (image) field to the Institution model (!159)
 * feat: create permissions class to check caregiver-patient access (!169)
 * Adjust legacy chart api endpoint to use legacy_id (!185)
+* feat: add model for a user's mobile device (!187)
 
 ### Documentation